+++ src/gmail/drafts.vue
@@ -12,7 +12,7 @@
           i more_vert
     .col
       nav.grey-text.right-align
-        label from 1 to 1
+        label from 1 to {{ emails.length }}
         a
           i arrow_back
         a
@@ -37,33 +37,12 @@
 </template>
 
 <script>
+import domain from "./domain";
+
 export default {
-  data() {
-    return {
-      check: false,
-      emails: [],
-    };
-  },
-  watch: {
-    check() {
-      this.checkAll();
-    },
-  },
-  mounted() {
-    for (var i = 0; i < 1; i++)
-      this.emails.push({ check: false, star: false, read: false });
-  },
-  methods: {
-    checkAll() {
-      for (var i = 0; i < this.emails.length; i++)
-        this.emails[i].check = this.check;
-    },
-    check(email) {
-      email.check = !email.check;
-    },
-    star(email) {
-      email.star = !email.star;
-    },
+  mixins: [domain],
+  created() {
+    this.emails = this.drafts;
   },
 };
 </script>