+++ src/uber/layout.vue
@@ -66,7 +66,7 @@ export default {
   },
   mounted() {
     this.updateTheme();
-    window.ui();
+    ui();
   },
   methods: {
     updateTheme() {