+++ src/youtube/whatsHot.vue
@@ -3,84 +3,23 @@
   .page.right.active
     h5 What's hot
     .row
-      .col.s12
+      .col.s12(v-for="item in itens")
         .card.no-padding.flat.transparent
           .row
             .col.s6.m4.l3
               a.wave
-                img.empty-state(:src="'/alok-001.jpg'")
+                img.empty-state(:src="item.image")
                 .absolute.right.bottom.small-margin.black.white-text.small-text &nbsp;00:00:00&nbsp;
             .col.s6.m8.l9
-              h5.no-margin Alok 01/2021
-              .grey-text 10k views
-              .space
-              .grey-text Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam eget ligula eu lectus lobortis condimentum. Aliquam nonummy auctor massa. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas.
-      .col.s12
-        .card.no-padding.flat.transparent
-          .row
-            .col.s6.m4.l3
-              a.wave
-                img.empty-state(:src="'/alok-002.jpg'")
-                .absolute.right.bottom.small-margin.black.white-text.small-text &nbsp;00:00:00&nbsp;
-            .col.s6.m8.l9
-              h5.no-margin Alok 12/2020
-              .grey-text 10k views
-              .space
-              .grey-text Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam eget ligula eu lectus lobortis condimentum. Aliquam nonummy auctor massa. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas.
-      .col.s12
-        .card.no-padding.flat.transparent
-          .row
-            .col.s6.m4.l3
-              a.wave
-                img.empty-state(:src="'/vintage-001.jpg'")
-                .absolute.right.bottom.small-margin.black.white-text.small-text &nbsp;00:00:00&nbsp;
-            .col.s6.m8.l9
-              h5.no-margin Best of 2021
-              .grey-text 10k views
-              .space
-              .grey-text Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam eget ligula eu lectus lobortis condimentum. Aliquam nonummy auctor massa. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas.
-      .col.s12
-        .card.no-padding.flat.transparent
-          .row
-            .col.s6.m4.l3
-              a.wave
-                img.empty-state(:src="'/vintage-002.jpg'")
-                .absolute.right.bottom.small-margin.black.white-text.small-text &nbsp;00:00:00&nbsp;
-            .col.s6.m8.l9
-              h5.no-margin Best of 2020
-              .grey-text 10k views
-              .space
-              .grey-text Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam eget ligula eu lectus lobortis condimentum. Aliquam nonummy auctor massa. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas.
-      .col.s12
-        .card.no-padding.flat.transparent
-          .row
-            .col.s6.m4.l3
-              a.wave
-                img.empty-state(:src="'/radio-001.jpg'")
-                .absolute.right.bottom.small-margin.black.white-text.small-text &nbsp;00:00:00&nbsp;
-            .col.s6.m8.l9
-              h5.no-margin Radio online
-              .grey-text 10k views
-              .space
-              .grey-text Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam eget ligula eu lectus lobortis condimentum. Aliquam nonummy auctor massa. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas.
-      .col.s12
-        .card.no-padding.flat.transparent
-          .row
-            .col.s6.m4.l3
-              a.wave
-                img.empty-state(:src="'/ocean-001.jpg'")
-                .absolute.right.bottom.small-margin.black.white-text.small-text &nbsp;00:00:00&nbsp;
-            .col.s6.m8.l9
-              h5.no-margin Alok, Zebra, Iro - Ocean
+              h5.no-margin {{ item.title }}
               .grey-text 10k views
               .space
               .grey-text Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam eget ligula eu lectus lobortis condimentum. Aliquam nonummy auctor massa. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas.
 </template>
 
 <script>
+import domain from "./domain";
 export default {
-  data() {
-    return;
-  },
+  mixins: [domain],
 };
-</script>s
+</script>